requirements.txt
# Core RAG dependencies - Latest stable versions
langchain>=0.2.0,<0.4.0
langchain-openai>=0.1.0,<0.3.0
langchain-community>=0.2.0,<0.4.0

# Vector databases
chromadb>=0.4.22,<0.6.0
faiss-cpu>=1.7.4,<1.9.0

# Text processing
tiktoken>=0.5.2,<0.8.0
sentence-transformers>=2.7.0,<3.0.0

# Korean language support
kiwipiepy>=0.15.2,<0.17.0
konlpy>=0.6.0,<0.7.0

# Data handling
pandas>=2.1.4,<3.0.0
numpy>=1.24.3,<2.0.0

# Environment management
python-dotenv>=1.0.0,<2.0.0

# Web framework (for API)
fastapi>=0.104.1,<0.115.0
uvicorn>=0.24.0,<0.32.0

# Development tools
jupyter>=1.0.0,<2.0.0

# Messaging system
redis>=4.5.0,<6.0.0
pydantic>=2.5.0,<3.0.0

# Async support
asyncio-mqtt>=0.16.0,<0.17.0  # Alternative messaging option
